Update automatic install instructions.

+++ install.sh
@@ -3,15 +3,15 @@
 cd
 
 # Install dependencies
-apt-get install -y git python-pip exuberant-ctags
-pip install flake8 jsbeautifier vim-debug dbgp
+sudo apt-get install -y git python-pip exuberant-ctags
+pip install --user flake8 jsbeautifier vim-debug dbgp
 
 # In case that you have existing vim configurations
-mkdir ~/.vim.old/
-mv .vim .vimrc .viminfo .vim.old/
-rmdir ~/.vim.old/  # this will remove the dir if no file was moved into
+if [[ -d ~/.vim/ || -f .vimrc || -f .viminfo ]]; then
+    mkdir ~/.vim.old/
+    mv .vim .vimrc .viminfo .vim.old/
+fi
 
 # Clone repo and symlink the vimrc file
 git clone https://github.com/ivanalejandro0/vim-configs.git .vim
 ln -s .vim/dot.vimrc .vimrc
-
